requirements.txt
streamlit
google-generativeai
pandas
plotly
st-supabase-connection